Modify Books.js file
