+++ src/components/Books.js
@@ -1,25 +1,12 @@
 import React from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import { useSelector } from 'react-redux';
 import booksSelector from '../redux/books/booksSelector';
 import SingleBook from './SingleBook';
 import AddBook from './AddBook';
-import { createBook } from '../redux/books/books_creators';
 
 const Books = () => {
-  const dispatch = useDispatch();
   const booksArray = useSelector(booksSelector);
-  const books = booksArray.map((book) => <SingleBook book={book} key={book.id} />);
-
-  const addBookToStore = (book) => {
-    const newBook = {
-      ...book,
-      author: 'Author',
-      percentProgress: '0',
-      currentChapter: 'Introduction',
-      chapterTitle: '',
-    };
-    dispatch(createBook(newBook));
-  };
+  const books = booksArray.map((book) => <SingleBook book={book} key={book.item_id} />);
 
   return (
     <div className="Books">
@@ -27,7 +14,7 @@ const Books = () => {
         { books }
       </div>
       <hr />
-      <AddBook addBookToStore={addBookToStore} />
+      <AddBook />
     </div>
   );
 };
